CoverageDiscarded to NotCoveredFrames
